+++ react/components/QuantitySelector/index.js
@@ -1,10 +1,24 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
+import DecreaseIcon from './images/DecreaseIcon'
+import Plus from '@vtex/styleguide/lib/icon/Plus'
+import { debounce } from 'underscore'
+
+import './global.css'
+
+const iconColor = '#C4C4C4'
 
 /**
  * Quantity selector component.
  */
 class QuantitySelector extends Component {
+  constructor(props) {
+    super(props)
+    this.debouncedDecreaseFunction = debounce(() => this.decreaseQuantity(), 1000)
+    this.debouncedIncreaseFunction = debounce(() => this.increaseQuantity(), 1000)
+    this.state = { currentQuantity: props.currentQuantity }
+  }
+
   handleChange = event => {
     const { maxQuantity, onQuantityChange, onMaxReached } = this.props
     const value = event.target.value
@@ -13,21 +27,63 @@ class QuantitySelector extends Component {
       quantity = maxQuantity
       onMaxReached()
     } else if (quantity < 0) {
-      quantity = 0
+      quantity = 1
     }
     onQuantityChange(quantity)
   }
 
+  handleDecreaseButtonClick = (event) => {
+    const { currentQuantity } = this.state
+    if (currentQuantity > 1) {
+      this.setState({ currentQuantity: currentQuantity - 1 })
+      event.persist()
+      this.debouncedDecreaseFunction()
+    }
+  }
+
+  handleIncreaseButtonClick = (event) => {
+    const { maxQuantity, onMaxReached } = this.props
+    const { currentQuantity } = this.state
+    if (currentQuantity < maxQuantity) {
+      this.setState({ currentQuantity: currentQuantity + 1 })
+      event.persist()
+      this.debouncedIncreaseFunction()
+    } else if (onMaxReached) {
+      onMaxReached()
+    }
+  }
+
+  decreaseQuantity = () => {
+    const { onQuantityChange } = this.props
+    const { currentQuantity } = this.state
+    onQuantityChange(currentQuantity)
+  }
+
+  increaseQuantity = () => {
+    const { currentQuantity } = this.state
+    const { onQuantityChange } = this.props
+    onQuantityChange(currentQuantity)
+  }
+
   render() {
-    const { currentQuantity } = this.props
+    const { currentQuantity } = this.state
+
     return (
-      <input
-        className="vtex-quantity-selector f4 dark-gray br2-l o-100"
-        type="number"
-        name="quantity"
-        value={currentQuantity}
-        onChange={this.handleChange}
-      />
+      <div className="flex flex-row">
+        <div className="pointer flex items-center justify-center" onClick={(e) => this.handleDecreaseButtonClick(e)}>
+          <DecreaseIcon />
+        </div>
+        <input
+          className="vtex-quantity-selector__input ma0 mh1 border-box bw1 br2 b--solid outline-0 near-black b--light-gray hover-b--silver bg-white f6 tc"
+          type="number"
+          name="quantity"
+          value={currentQuantity}
+          onChange={this.handleChange}
+        />
+        <div className="pointer flex items-center justify-center" onClick={(e) => this.handleIncreaseButtonClick(e)}>
+          <Plus color={iconColor} />
+        </div>
+      </div>
     )
   }
 }
@@ -40,7 +96,11 @@ QuantitySelector.propTypes = {
   /** Called when the client set the quantity selector */
   onQuantityChange: PropTypes.func.isRequired,
   /** Define if can buy more items than the maximum limit */
-  onMaxReached: PropTypes.func.isRequired,
+  onMaxReached: PropTypes.func,
+}
+
+QuantitySelector.defaultProps = {
+  currentQuantity: 0,
 }
 
 export default QuantitySelector