+++ react/components/SKUSelector/components/SelectorItem.js
@@ -1,12 +1,13 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
+import { FormattedNumber } from 'react-intl'
 
 import VTEXClasses from '../constants/CustomClasses'
 
 /**
  * Inherits the components that should be displayed inside the Selector component.
  */
-class SelectorItem extends PureComponent {
+export default class SelectorItem extends PureComponent {
   handleClick = event => {
     event.preventDefault()
     if (this.props.onClick) {
@@ -15,9 +16,18 @@ class SelectorItem extends PureComponent {
   }
 
   render() {
+    const { isAvailable, isSelected, children, maxPrice, price } = this.props
+    const discount = getDiscount(maxPrice, price)
     return (
-      <div className={`${VTEXClasses.SELECTOR__ITEM} di ba bw1 pointer ${this.props.isSelected ? 'b--black' : 'b--transparent'}`} onClick={this.handleClick}>
-        { this.props.children }
+      <div
+        className={`${VTEXClasses.SELECTOR__ITEM} di ba bw1 pointer flex items-center relative
+        ${isSelected ? 'b--blue' : 'b--transparent'}
+        ${!isAvailable && 'bg-light-gray'}`}
+        onClick={this.handleClick}>
+        <div className={`${!isAvailable && 'o-50'}`}>
+          {children}
+        </div>
+        {discount > 0 && <span className={`${VTEXClasses.SKU_BADGE} b`}><FormattedNumber value={discount} style="percent" /></span>}
       </div>
     )
   }
@@ -30,11 +40,27 @@ SelectorItem.propTypes = {
   children: PropTypes.node,
   /** Function that is called when the item is clicked */
   onClick: PropTypes.func,
+  /** Flag that indicates if the sku is available */
+  isAvailable: PropTypes.bool,
+  /** Flag that indicates if the current item is selected */
+  isSelected: PropTypes.bool,
+  /** Max sku price */
+  maxPrice: PropTypes.number,
+  /** Price of the current sku */
+  price: PropTypes.number,
 }
 
 SelectorItem.defaultProps = {
   index: 0,
   children: {},
+  isAvailable: true,
+  isSelected: false,
 }
 
-export default SelectorItem
+const getDiscount = (maxPrice, price) => {
+  let discount = 0
+  if (maxPrice && price) {
+    discount = 1 - (price / maxPrice)
+  }
+  return discount
+}